chore: refactor main.rs entry to make test easily
