+++ README.md
@@ -166,4 +166,16 @@ curl -fsSL https://raw.githubusercontent.com/sheinsight/snm/main/install.sh | ba
 - [x] check valid package manager
 - [x] execute package manager command
 
- 
+
+
+ ### Run test
+
+use [cargo-nextest](https://nexte.st/book/pre-built-binaries) to run tests
+
+for Mac:
+
+ ```
+ curl -LsSf https://get.nexte.st/latest/mac | tar zxf - -C ${CARGO_HOME:-~/.cargo}/bin
+
+ cargo nextest run
+ ```
