+++ crates/cli/tests/no_package_manager_property/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "no_node_version",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
