+++ crates/cli/tests/tests.rs
@@ -1,0 +1,5 @@
+#[path = "no_package_manager_property/no_package_manager_property_test.rs"]
+mod no_package_manager_property_test;
+
+#[path = "no_package_json/no_package_json_test.rs"]
+mod no_package_json_test;
