+++ crates/snm_core/tests/tests.rs
@@ -1,0 +1,1 @@
+mod snm_config_init_test;
